# internal/apidoc/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: Aegis
    url: https://github.com/wisonwang/Aegis
  description: Aegis — AI Data Supply Gateway. 受治理的 DataAPI + 面向 AI Agent 的 MCP 服务。默认拒绝：每条查询都经过权限引擎重写与脱敏。
  license:
    name: MIT
  termsOfService: https://github.com/wisonwang/Aegis
  title: Aegis DataAPI
  version: "1.0"
paths:
...
  /admin/api/workspaces:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Create Workspace
      tags:
      - workspaces
  /admin/api/workspaces/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Workspace
      tags:
      - workspaces
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Get Workspace
      tags:
      - workspaces
  /admin/api/workspaces/{id}/members:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Workspace Members
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Add Workspace Member
      tags:
      - workspaces
  /admin/api/workspaces/{id}/members/{user_id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: user_id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Remove Workspace Member
      tags:
      - workspaces
...
  /api/v1/workspaces:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: list My Workspaces
      tags:
      - workspaces
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT> — 通过 POST /api/v1/login 获取"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"